requirements.txt
pyTelegramBotAPI==4.23.0
psycopg2==2.9.9
icecream==2.1.3
python-dotenv==1.0.1
requests==2.32.3